Add full packet capture of macOS and Touch Bar
The added file contains a full packet capture of a VM booting macOS
and properly detecting the Touch Bar. During this capture Safari and
the "Control Strip settings" got opened to get packets of macOS using
the advanced capabilities of using the Touch Bar.

What didn't work in this setup however was input via the Touch Bar.
Input event simply didn't register, except for the "Done" button in
the "Control Strip" configuration.